added debug mode to trainer, minor changes
